feat(about): Update drug resistance definitions
Updated the drug resistance definitions in the About component to reflect the latest information. The changes include updates to the names and criteria for drug resistance determinants such as Ciprofloxacin non-susceptibility (CipNS/R), Third-generation cephalosporin resistance (3 GCR), and Azithromycin resistance (AzithR). The definitions now align with the referenced source, Argimon et al, 2021.

fix(download): Correct typo in time period

Fixed a typo in the DownloadData component where the label "Time Period" was misspelled as "Time Period". The label has been corrected to "Time period" for consistency.

style(graphs): Update button label

Updated the label of a button in the FrequenciesGraph component from "Clear All" to "Clear" for brevity and clarity.

style(map): Update labels for time period

Updated the labels in the BottomRightControls and TopRightControls2 components to use "Time period" instead of "Time Period" for consistency.

style(graphColorHelper): Modify comment and drug name

Modified a comment in the graphColorHelper file and changed the drug name from "Fluoroquinolones (CipI)" to "